+++ src/main/java/com/ziggs/ziggs_backend/repository/SoundDataRepository.java
@@ -3,5 +3,5 @@ package com.ziggs.ziggs_backend.repository;
 import com.ziggs.ziggs_backend.entity.SoundData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SoundDataRepository extends JpaRepository<SoundData, Integer> {
+public interface SoundDataRepository extends JpaRepository<SoundData, Long> {
 }
